// apps/frontend/src/types/prontuario.ts
export interface Prontuario {
  id: string;
  pacienteId: string;
  pacienteNome: string;
  data: string;
  profissional: string;
  queixaPrincipal: string;
  objetivosSessao: string;
  atividadesRealizadas: string;
  resultados: string;
  comportamento: string;
  orientacoesPais: string;
  proximaMeta: string;
}